Rename GBs to GiB's for memory cost
There's a number of references in the exported metrics that state the hourly cost of memory and persistent volumes is in GB's. Azure, AWS, and GCP all have pricing models based upon gibibytes(GiB) and is causing some confusion about what the cost is reflective of. 

This updates all references of pricing to be in GiB. 

Signed-off-by: Mark <mark.poko@grafana.com>